feat: add MSL scalar typedefs (uchar/ushort/uint/ulong/half) and METAL_FUNC macro
Metal Shading Language requires these builtin typedefs as the Apple
standard library draws on them without including any header.  Without
them, every file in metal_stdlib that uses uchar, ushort, uint, ulong,
or half fails to parse.

Changes:
- Sema.cpp: inject implicit typedefs for uchar, ushort, uint, ulong,
  and half alongside the existing Metal opaque handle types.  The
  reference AST dumps confirm Apple's compiler does exactly this
  (TypedefDecl entries at the head of every -x metal translation unit).
- MetalMacros.def: add METAL_FUNC predefined macro (Apple's compiler
  defines METAL_FUNC as static __inline__ __attribute__((always_inline)))
- math-fast.metal, builtin-arity.metal, type-names.metal: use half
  (now available as builtin typedef) instead of __fp16 workaround
- run-tests.sh: use -isystem (not -I) for stdlib path to suppress
  warnings from Apple's own headers

+++ clang/test/Metal/type-names.metal
@@ -39,14 +39,14 @@ using namespace metal;
 //
 // CHECK-DAG: !"air.arg_type_name", !"texture2d<float, sample>"
 // CHECK-DAG: !"air.arg_type_name", !"texture2d<float, write>"
-// CHECK-DAG: !"air.arg_type_name", !"texture3d<float, read>"
+// CHECK-DAG: !"air.arg_type_name", !"texture3d<half, read>"
 //
 // Scalars use the MSL spelling; "unsigned int" must not appear.
 // CHECK-DAG: !"air.arg_type_name", !"uint"
 // CHECK-NOT: !"air.arg_type_name", !"unsigned int"
 kernel void k(texture2d<float> defaulted [[texture(0)]],
               texture2d<float, access::write> explicit_write [[texture(1)]],
-              texture3d<float> depth_defaulted [[texture(2)]],
+              texture3d<half> depth_defaulted [[texture(2)]],
               device uint *out [[buffer(0)]],
               uint id [[thread_position_in_grid]]) {
   out[id] = id;
